Added technical test project to new repository
Up to now. the new user form can be filled out and sent. I had an issue republishing a DB model after I changed some of the details in it so I can't currently get that to work. But elsewise, the form should be able to be submitted. The API side of the DB is in place also. I should be able to make a request to call a list of all users using the API.